requirements/common.txt
alembic==1.6.5
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
Flask==1.1.4
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==3.2
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==2.0.1
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.22
urllib3==1.26.6
Werkzeug==1.0.1
